use parsefloat and update db to store float

+++ app/src/app/api/tastings/[id]/ratings/route.ts
@@ -9,7 +9,7 @@ export const POST = async (
   const tastingId = params.id
   const body = await request.json()
   const userEmail = body.userEmail as string
-  const rating = parseInt(body.rating)
+  const rating = parseFloat(body.rating)
 
   const tasting = await prisma.tasting.findUnique({
     where: { id: tastingId },
